UCP: Move flush-related functions to a separate file.
