+++ src/ucp/core/ucp_ep.h
@@ -223,6 +223,10 @@ ucs_status_t ucp_ep_new(ucp_worker_h worker, uint64_t dest_uuid,
                         const char *peer_name, const char *message,
                         ucp_ep_h *ep_p);
 
+ucs_status_ptr_t ucp_disconnect_nb_internal(ucp_ep_h ep, unsigned mode);
+
+void ucp_ep_disconnected(ucp_ep_h ep);
+
 ucs_status_t ucp_ep_create_stub(ucp_worker_h worker, uint64_t dest_uuid,
                                 const char *message, ucp_ep_h *ep_p);
