+++ src/ucp/rma/basic_rma.c
@@ -420,53 +420,3 @@ out:
     return status;
 }
 
-UCS_PROFILE_FUNC(ucs_status_t, ucp_worker_flush, (worker), ucp_worker_h worker)
-{
-    unsigned rsc_index;
-
-    UCP_THREAD_CS_ENTER_CONDITIONAL(&worker->mt_lock);
-
-    while (worker->wireup_pend_count > 0) {
-        ucp_worker_progress(worker);
-    }
-
-    /* TODO flush in parallel */
-    for (rsc_index = 0; rsc_index < worker->context->num_tls; ++rsc_index) {
-        if (worker->ifaces[rsc_index].iface == NULL) {
-            continue;
-        }
-
-        while (uct_iface_flush(worker->ifaces[rsc_index].iface, 0, NULL) != UCS_OK) {
-            ucp_worker_progress(worker);
-        }
-    }
-
-    UCP_THREAD_CS_EXIT_CONDITIONAL(&worker->mt_lock);
-
-    return UCS_OK;
-}
-
-UCS_PROFILE_FUNC(ucs_status_t, ucp_ep_flush, (ep), ucp_ep_h ep)
-{
-    ucp_lane_index_t lane;
-    ucs_status_t     status;
-
-    UCP_THREAD_CS_ENTER_CONDITIONAL(&ep->worker->mt_lock);
-
-    for (lane = 0; lane < ucp_ep_num_lanes(ep); ++lane) {
-        for (;;) {
-            status = uct_ep_flush(ep->uct_eps[lane], 0, NULL);
-            if (status == UCS_OK) {
-                break;
-            } else if ((status != UCS_INPROGRESS) && (status != UCS_ERR_NO_RESOURCE)) {
-                goto out;
-            }
-            ucp_worker_progress(ep->worker);
-        }
-    }
-
-    status = UCS_OK;
-out:
-    UCP_THREAD_CS_EXIT_CONDITIONAL(&ep->worker->mt_lock);
-    return status;
-}
